requirements.txt
streamlit
yfinance
elevenlabs
python-dotenv
plotly
